Parameter studies output overwriting (#52)
* add delete output parameter

* submodule update

+++ SPHINXsys/src/shared/io_system/in_output.cpp
@@ -11,7 +11,7 @@
 namespace SPH
 {
 	//=============================================================================================//
-	In_Output::In_Output(SPHSystem& sph_system)
+	In_Output::In_Output(SPHSystem& sph_system, bool delete_output)
 		: sph_system_(sph_system),
 		input_folder_("./input"), output_folder_("./output"),
 		restart_folder_("./restart"), reload_folder_("./reload")
@@ -35,9 +35,11 @@ namespace SPH
 		{
 			fs::remove_all(restart_folder_);
 			fs::create_directory(restart_folder_);
-
-			fs::remove_all(output_folder_);
-			fs::create_directory(output_folder_);
+			if(delete_output == true)
+			{
+				fs::remove_all(output_folder_);
+				fs::create_directory(output_folder_);
+			}
 		}
 
 		restart_step_ = std::to_string(sph_system.restart_step_);
